linux_fix

+++ include/res.hpp
@@ -2,6 +2,6 @@
 
 namespace res {
     void* get_asset_io(const char* fn);
-    void* read_asset_data(const char* fn, size_t& size_buf);
+    void* read_asset_data(const char* fn, int& size_buf);
     void free_asset_data(void* data);
 }
